azure-identity-blob/pyproject.toml
[tool.poetry]
name = "azure-identity-blob"
version = "0.1.0"
description = ""
authors = ["Katz Sakai"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
azure-storage-blob = "^12.8.1"
azure-identity = "^1.5.0"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
mypy = "^0.812"
black = "^21.5b0"
pytest = "^6.2.4"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
